refactor: split out logic into three files
